jlcparts_db_convert.py - Fix script to error out if the database is not present
Previously the script would create the database if it didn't exist, which doesn't make
sense as the database existing is a necessary prerequisite of the script.

This results in incorrect behavior if you run the script from a folder that
doesn't contain a db_build/cache.sqlite3 database file.

Use a uri string so sqlite3 will error if the database file does not exist

From https://stackoverflow.com/a/57828452/299037