Default to database priority
Defaulting the setting to board priority means that by default all
work done will be lost if you forget to click the export button.
This is dangerous and confusing, so instead this saves work by default.

It would be better to also warn when the dialog is closed without
exporting, when the setting is the other way.