Show component values in list (+ other minor conveniences) (#551)
* Make footprint table scrollable

* Make part details dialog resizable

* Default to database priority

Defaulting the setting to board priority means that by default all
work done will be lost if you forget to click the export button.
This is dangerous and confusing, so instead this saves work by default.

It would be better to also warn when the dialog is closed without
exporting, when the setting is the other way.

* Don't erase non-matching board parts

If in database priority mode, set the board's part to the
database's part rather than erasing it.

* Show parameters derived from LCSC description

This adds an "LCSC Params" field to the footprint and part selection lists
that makes it much easier to double-check the assignments or select the
right part. Particularly for passives, LCSC buries the important data (like
the resistance of a resistor) in the middle of the description field. This
pulls those values out where you can see them.

* Add double-click behaviors

Double-clicking a footprint brings up part selection.
Double-clicking a part in the selector assigns the part.

* Improve default search string

Don't search for the existing LCSC value -- if I'm searching, that's
presumably the only part I don't want!

Tweak the search string to make it easier to find passives quickly.

* placehoDLer -> placehoLDer (#553)

Fix for little typo

* ruff auto fix errors

* Fix remaning ruff erorrs

---------

Co-authored-by: Maxime Vincent <maxime@veemax.be>
Co-authored-by: bouni <bouni@owee.de>